Begur updates

+++ be/tba/ejb/task/session/TaskSqlAdapter.java
@@ -1,11 +1,14 @@
 package be.tba.ejb.task.session;
 
+import java.util.Calendar;
 import java.util.Collection;
 import java.util.Hashtable;
 import java.util.Iterator;
 import java.util.Vector;
 
 import be.tba.ejb.account.interfaces.AccountEntityData;
+import be.tba.ejb.invoice.interfaces.InvoiceEntityData;
+import be.tba.ejb.invoice.session.InvoiceSqlAdapter;
 import be.tba.ejb.task.interfaces.TaskEntityData;
 import be.tba.servlets.session.WebSession;
 import be.tba.util.data.AbstractSqlAdapter;
@@ -101,12 +104,12 @@ public class TaskSqlAdapter extends AbstractSqlAdapter<TaskEntityData>
     /**
      * @ejb:interface-method view-type="remote"
      */
-    public Collection<TaskEntityData> getTasksFromTillTimestamp(WebSession webSession, String fwdNr, long start, long stop)
+    public Collection<TaskEntityData> getTasksFromTillTimestamp(WebSession webSession, int accountId, long start, long stop)
     {
         try
         {
             Collection<TaskEntityData> vTaskList = new Vector<TaskEntityData>();
-            collectInvoiceTasks(webSession, AccountCache.getInstance().get(fwdNr), vTaskList, start, stop);
+            collectInvoiceTasks(webSession, AccountCache.getInstance().get(accountId), vTaskList, start, stop);
             return vTaskList;
         }
         catch (Exception e)
@@ -119,12 +122,12 @@ public class TaskSqlAdapter extends AbstractSqlAdapter<TaskEntityData>
     /**
      * @ejb:interface-method view-type="remote"
      */
-    public Hashtable<String, Collection<TaskEntityData>> getTasksFromTillTimestampHashtable(WebSession webSession, String fwdNr, long start, long stop)
+    public Hashtable<Integer, Collection<TaskEntityData>> getTasksFromTillTimestampHashtable(WebSession webSession, int accountId, long start, long stop)
     {
         try
         {
-            Hashtable<String, Collection<TaskEntityData>> vTaskList = new Hashtable<String, Collection<TaskEntityData>>();
-            collectInvoiceTasksHashTable(webSession, AccountCache.getInstance().get(fwdNr), vTaskList, start, stop);
+            Hashtable<Integer, Collection<TaskEntityData>> vTaskList = new Hashtable<Integer, Collection<TaskEntityData>>();
+            collectInvoiceTasksHashTable(webSession, AccountCache.getInstance().get(accountId), vTaskList, start, stop);
 
             //System.out.println("getTasksFromTillTimestamp for " + fwdNr + ": " + vTaskList.size() + " entries.");
             return vTaskList;
@@ -133,15 +136,40 @@ public class TaskSqlAdapter extends AbstractSqlAdapter<TaskEntityData>
         {
             e.printStackTrace();
         }
-        return new Hashtable<String, Collection<TaskEntityData>>();
+        return new Hashtable<Integer, Collection<TaskEntityData>>();
     }
 
+    public Collection<TaskEntityData> getTasksForInvoice(WebSession webSession, int invoiceId)
+    {
+    	return executeSqlQuery(webSession, "SELECT * FROM TaskEntity WHERE InvoiceId=" + invoiceId + " ORDER BY TimeStamp DESC"); 
+    }
+
+    public void setInvoiceId(WebSession webSession, int key, int invoiceId)
+    {
+    	executeSqlQuery(webSession, "UPDATE TaskEntity set InvoiceId=" + invoiceId  + " where id=" + key); 
+    }
+    
+    public void fixDbIds(WebSession webSession, int key, int invoiceId, int accountId)
+    {
+    	executeSqlQuery(webSession, "UPDATE TaskEntity set InvoiceId=" + invoiceId + ", AccountID=" + accountId + " where id=" + key); 
+    }
+    
+    
+    
     /**
      * @ejb:interface-method view-type="remote"
      */
     public void removeTask(WebSession webSession, int key)
     {
-        executeSqlQuery(webSession, "DELETE FROM TaskEntity WHERE Id=" + key);
+    	TaskEntityData entry = getRow(webSession, key);
+    	if (entry != null && entry.getInvoiceId() == 0)
+    	{
+    		executeSqlQuery(webSession, "DELETE FROM TaskEntity WHERE Id=" + key);
+    	}
+    	else
+    	{
+    		System.out.println("task cannot be removed because it belongs to an invoice");
+    	}
     }
 
     public TaskEntityData getTask(WebSession webSession, int key)
@@ -173,7 +201,12 @@ public class TaskSqlAdapter extends AbstractSqlAdapter<TaskEntityData>
 
     private void collectInvoiceTasks(WebSession webSession, AccountEntityData customer, Collection<TaskEntityData> taskList, long start, long stop)
     {
-        Collection<TaskEntityData> vCollection =  queryAllTasksForFwdNr(webSession, customer.getFwdNumber(), start, stop);
+        if (customer == null)
+        {
+        	System.out.println("collectInvoiceTasks() with customer=null");
+    		return;
+        }
+    	Collection<TaskEntityData> vCollection =  queryAllTasksForFwdNr(webSession, customer.getFwdNumber(), start, stop);
         
         if (vCollection != null)
         {
@@ -181,7 +214,7 @@ public class TaskSqlAdapter extends AbstractSqlAdapter<TaskEntityData>
         }
         if (customer.getHasSubCustomers())
         {
-            Collection<AccountEntityData> vSubCustomerList = AccountCache.getInstance().getSubCustomersList(customer.getFwdNumber());
+            Collection<AccountEntityData> vSubCustomerList = AccountCache.getInstance().getSubCustomersList(customer.getId());
             //System.out.println(customer.getFullName() + "has " + vSubCustomerList.size() + " sub customers");
             for (Iterator<AccountEntityData> i = vSubCustomerList.iterator(); i.hasNext();)
             {
@@ -200,16 +233,16 @@ public class TaskSqlAdapter extends AbstractSqlAdapter<TaskEntityData>
         }
     }
 
-    private void collectInvoiceTasksHashTable(WebSession webSession, AccountEntityData customer, Hashtable<String, Collection<TaskEntityData>> taskList, long start, long stop)
+    private void collectInvoiceTasksHashTable(WebSession webSession, AccountEntityData customer, Hashtable<Integer, Collection<TaskEntityData>> taskList, long start, long stop)
     {
         Collection<TaskEntityData> vCollection = queryAllTasksForFwdNr(webSession, customer.getFwdNumber(), start, stop);
         if (vCollection != null)
         {
-            taskList.put(customer.getFwdNumber(), vCollection);
+            taskList.put(customer.getId(), vCollection);
         }
         if (customer.getHasSubCustomers())
         {
-            Collection<AccountEntityData> vSubCustomerList = AccountCache.getInstance().getSubCustomersList(customer.getFwdNumber());
+            Collection<AccountEntityData> vSubCustomerList = AccountCache.getInstance().getSubCustomersList(customer.getId());
             //System.out.println(customer.getFullName() + "has " + vSubCustomerList.size() + " sub customers");
             for (Iterator<AccountEntityData> i = vSubCustomerList.iterator(); i.hasNext();)
             {
@@ -220,7 +253,7 @@ public class TaskSqlAdapter extends AbstractSqlAdapter<TaskEntityData>
                     vCollection = queryAllTasksForFwdNr(webSession, vEntry.getFwdNumber(), start, stop);
                     if (vCollection != null)
                     {
-                        taskList.put(vEntry.getFwdNumber(), vCollection);
+                        taskList.put(vEntry.getId(), vCollection);
                     }
                 }
             }
@@ -231,6 +264,49 @@ public class TaskSqlAdapter extends AbstractSqlAdapter<TaskEntityData>
     {
         Collection<TaskEntityData> vCollection = executeSqlQuery(webSession, "SELECT * FROM TaskEntity WHERE FwdNr='" + fwdNr + "' AND TimeStamp>" + start + " AND TimeStamp<=" + stop + " AND IsRecuring=FALSE ORDER BY TimeStamp DESC");
         Collection<TaskEntityData> vRecuringCollection = executeSqlQuery(webSession, "SELECT * FROM TaskEntity WHERE FwdNr='" + fwdNr + "' AND StartTime<" + stop + " AND IsRecuring=TRUE ORDER BY TimeStamp DESC");
+        if (vCollection != null)
+        {
+            vCollection.addAll(vRecuringCollection);
+        }
+        else
+        {
+            vCollection = vRecuringCollection;
+        }
+        return vCollection;
+    }
+    
+    private Collection<TaskEntityData> queryNotInvoicedTasksForFwdNr(WebSession webSession, int accountId)
+    {
+    	Collection<TaskEntityData> vRecuringCollection = new Vector<TaskEntityData>();
+    	Collection<TaskEntityData> vCollection = executeSqlQuery(webSession, "SELECT * FROM TaskEntity WHERE AccountID='" + accountId + "' AND IsInvoiced=false AND IsRecuring=FALSE ORDER BY TimeStamp DESC");
+        
+        InvoiceSqlAdapter vInvoiceSession = new InvoiceSqlAdapter();
+        Calendar vStartCalendar = Calendar.getInstance();
+        int year = vStartCalendar.get(Calendar.YEAR);
+        int month = vStartCalendar.get(Calendar.MONTH);
+        
+        boolean isRecuringAlreadyInvoiced = false;
+        
+        // whether this recuring task has already been invoiced this month
+        Collection<InvoiceEntityData> invoices = vInvoiceSession.getInvoiceList(webSession, accountId, month, year);
+        if (!invoices.isEmpty())
+        {
+        	for (Iterator<InvoiceEntityData> iter = invoices.iterator(); iter.hasNext();)
+            {
+        		InvoiceEntityData vEntry = iter.next();
+        		if (vEntry.getFrozenFlag())
+        		{
+        			isRecuringAlreadyInvoiced = true;
+        			break;
+        		}
+            }
+        }
+        	
+        if (!isRecuringAlreadyInvoiced)	
+        {
+        	long now = Calendar.getInstance().getTimeInMillis();
+            vRecuringCollection = executeSqlQuery(webSession, "SELECT * FROM TaskEntity WHERE AccountID='" + accountId + "' AND StartTime<" + now + "' AND StopTime>" + now+ " AND IsRecuring=TRUE ORDER BY TimeStamp DESC");
+        }
         if (vCollection != null)
         {
             vCollection.addAll(vRecuringCollection);
@@ -271,11 +347,12 @@ public class TaskSqlAdapter extends AbstractSqlAdapter<TaskEntityData>
             entry.setFixedPrice(rs.getDouble(6));
             entry.setTimeSpend(rs.getInt(7));
             entry.setDescription(null2EmpthyString(rs.getString(8)));
-            entry.setIsInvoiced(rs.getBoolean(9));
-            entry.setIsRecuring(rs.getBoolean(10));
-            entry.setStartTime(rs.getLong(11));
-            entry.setStopTime(rs.getLong(12));
-            entry.setDoneBy(null2EmpthyString(rs.getString(13)));
+            entry.setIsRecuring(rs.getBoolean(9));
+            entry.setStartTime(rs.getLong(10));
+            entry.setStopTime(rs.getLong(11));
+            entry.setDoneBy(null2EmpthyString(rs.getString(12)));
+            entry.setInvoiceId(rs.getInt(13));
+            entry.setAccountId(rs.getInt(14));
             vVector.add(entry);
         }
         return vVector;
